031_Project/012_bonus/requirements.txt
beautifulsoup4==4.11.1
matplotlib==3.6.2
hvplot==0.8.2
pandas==1.5.2
plotly==5.11.0
python-slugify==7.0.0
